+++ src/Microsoft.ApplicationInsights/Extensibility/Implementation/RichPayloadEventSource.cs
@@ -60,6 +60,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.Requests);
             }
             else if (item is TraceTelemetry)
@@ -76,6 +77,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.Traces);
             }
             else if (item is EventTelemetry)
@@ -92,6 +94,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.Events);
             }
             else if (item is DependencyTelemetry)
@@ -108,6 +111,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.InternalData,
+                    telemetryItem.Context.Flags,
                     Keywords.Dependencies);
             }
             else if (item is MetricTelemetry)
@@ -124,6 +128,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.Metrics);
             }
             else if (item is ExceptionTelemetry)
@@ -140,6 +145,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.Exceptions);
             }
 #pragma warning disable 618
@@ -157,6 +163,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.Metrics);
             }
 #pragma warning restore 618
@@ -174,6 +181,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.PageViews);
             }
 #pragma warning disable 618
@@ -191,6 +199,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.Events);
             }
             else if (item is AvailabilityTelemetry)
@@ -207,6 +216,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.SanitizedTags,
                     telemetryItem.Data,
+                    telemetryItem.Context.Flags,
                     Keywords.Availability);
             }
             else
@@ -264,12 +274,12 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation
             }
         }
 
-        private void WriteEvent<T>(string eventName, string instrumentationKey, IDictionary<string, string> tags, T data, EventKeywords keywords)
+        private void WriteEvent<T>(string eventName, string instrumentationKey, IDictionary<string, string> tags, T data, long flags, EventKeywords keywords)
         {
             this.EventSourceInternal.Write(
                 eventName,
                 new EventSourceOptions() { Keywords = keywords },
-                new { PartA_iKey = instrumentationKey, PartA_Tags = tags, _B = data });
+                new { PartA_iKey = instrumentationKey, PartA_Tags = tags, _B = data, PartA_flags = flags });
         }
 
         private void WriteEvent(OperationTelemetry item, EventOpcode eventOpCode)